requirements.txt
pandas>=2.0,<3.0
pytest>=7.0
python-dotenv>=1.0
